requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
ollama==0.4.7
openai==1.68.2
pip==25.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyperclip==1.9.0
python-dotenv==1.0.1
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
